requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

cycler==0.11.0; python_version >= '3.6'

fonttools==4.32.0; python_version >= '3.7'

imutils==0.5.4

kiwisolver==1.4.2; python_version >= '3.7'

matplotlib==3.5.1

numpy==1.22.3

opencv-python==4.5.5.64

packaging==21.3; python_version >= '3.6'

pillow==9.1.0

pyparsing==3.0.8; python_full_version >= '3.6.8'

pytesseract==0.3.9

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

